Move: `diff_manifests`, `manifest_metaspec` -> manio.py
