+++ test_manio.py
@@ -6,6 +6,7 @@ import square
 from test_helpers import make_manifest, mk_deploy
 
 import unittest.mock as mock
+import k8s_utils
 
 
 # Convenience.
@@ -525,6 +526,165 @@ class TestYamlManifestIO:
         }
         expected = self.yamlfy(expected)
         assert fdata_raw_new == expected
+
+
+class TestManifestValidation:
+    @classmethod
+    def setup_class(cls):
+        square.setup_logging(9)
+
+    def test_manifest_metaspec_basic_valid(self):
+        """Ensure it returns only the salient fields of valid manifests."""
+
+        # Minimal Deployment manifest with just the salient fields. Test
+        # function must return a deepcopy of it.
+        valid_deployment_manifest = {
+            'apiVersion': 'v1',
+            'kind': 'Deployment',
+            'metadata': {'name': 'foo', 'namespace': 'bar'},
+            'spec': {'some': 'thing'},
+        }
+        ret = manio.manifest_metaspec(valid_deployment_manifest)
+        assert ret == RetVal(valid_deployment_manifest, False)
+        assert valid_deployment_manifest is not ret.data
+
+        # Minimal Namespace manifest with just the salient fields. Test
+        # function must return a deepcopy of it.
+        valid_namespace_manifest = {
+            'apiVersion': 'v1',
+            'kind': 'Namespace',
+            'metadata': {'name': 'foo'},
+            'spec': {'some': 'thing'},
+        }
+        ret = manio.manifest_metaspec(valid_namespace_manifest)
+        assert ret == RetVal(valid_namespace_manifest, False)
+        assert valid_namespace_manifest is not ret.data
+
+        # Function must accept additional entries (eg "additional" in example
+        # below) but not return them. It must not matter whether those
+        # additional entries are actually valid keys in the manifest.
+        valid_namespace_manifest_add = {
+            'apiVersion': 'v1',
+            'kind': 'Namespace',
+            'metadata': {'name': 'foo', 'additional': 'entry'},
+            'spec': {'some': 'thing'},
+            'status': {"some": "status"},
+            'additional': 'entry',
+        }
+        ret = manio.manifest_metaspec(valid_namespace_manifest_add)
+        assert ret == RetVal(valid_namespace_manifest, False)
+
+    def test_manifest_metaspec_automanifests(self):
+        """Verify that it works with the `make_manifest` test function.
+
+        This test merely validates that the output of the `make_manifest`
+        function used in various tests produces valid manifests as far as
+        `manifest_metaspec` is concerned.
+
+        """
+        # Create a valid manifest for each supported resource kind and verify
+        # that the test function accepts it.
+        for kind in square.SUPPORTED_KINDS:
+            if kind == "Namespace":
+                manifest = make_manifest(kind, None, "name")
+            else:
+                manifest = make_manifest(kind, "ns", "name")
+
+            ret = manio.manifest_metaspec(manifest)
+            assert ret.err is False and len(ret.data) > 0
+
+        # Invalid Namespace manifest: metadata.namespace field is not None.
+        manifest = make_manifest("Namespace", "ns", "name")
+        assert manio.manifest_metaspec(manifest) == RetVal(None, True)
+
+        # Unknown resource kind "foo".
+        manifest = make_manifest("foo", "ns", "name")
+        assert manio.manifest_metaspec(manifest) == RetVal(None, True)
+
+    def test_manifest_metaspec_missing_fields(self):
+        """Incomplete manifests must be rejected."""
+        # A valid deployment manifest.
+        valid = {
+            "apiVersion": "v1",
+            "kind": "Deployment",
+            "metadata": {"name": "foo", "namespace": "bar"},
+            "spec": {"some": "thing"},
+        }
+
+        # Create stunted manifests by creating a copy of `valid` that misses
+        # one key in each iteration. The test function must reject all those
+        # manifests and return an error.
+        for field in valid:
+            invalid = {k: v for k, v in valid.items() if k != field}
+            assert manio.manifest_metaspec(invalid) == RetVal(None, True)
+
+        # Metadata for Namespace manifests must contain a "name" field.
+        invalid = {
+            "apiVersion": "v1",
+            "kind": "Namespace",
+            "metadata": {"foo": "bar"},
+            "spec": {"some": "thing"},
+        }
+        assert manio.manifest_metaspec(invalid) == RetVal(None, True)
+
+        # Metadata for Namespace manifests must not contain a "namespace" field.
+        invalid = {
+            "apiVersion": "v1",
+            "kind": "Namespace",
+            "metadata": {"namespace": "namespace"},
+            "spec": {"some": "thing"},
+        }
+        assert manio.manifest_metaspec(invalid) == RetVal(None, True)
+
+        # Metadata for non-namespace manifests must contain "name" and "namespace".
+        invalid = {
+            "apiVersion": "v1",
+            "kind": "Deployment",
+            "metadata": {"name": "name"},
+            "spec": {"some": "thing"},
+        }
+        assert manio.manifest_metaspec(invalid) == RetVal(None, True)
+
+
+class TestDiff:
+    @classmethod
+    def setup_class(cls):
+        square.setup_logging(9)
+
+    def test_diff_manifests_ok(self):
+        """Diff two valid manifests and (roughly) verify the output."""
+        # Two valid manifests.
+        srv = make_manifest("Deployment", "namespace", "name1")
+        loc = make_manifest("Deployment", "namespace", "name2")
+
+        # Test function must able to cope with `DotDicts`.
+        srv = k8s_utils.make_dotdict(srv)
+        loc = k8s_utils.make_dotdict(loc)
+
+        # Diff the manifests. Must not return an error.
+        diff_str, err = manio.diff_manifests(loc, srv)
+        assert err is False
+
+        # Since it is difficult to compare the correct diff string due to
+        # formatting characters, we will only verify that the string contains
+        # a line that removes the old "names1" and one line to add "name2".
+        assert "-  name: name1" in diff_str
+        assert "+  name: name2" in diff_str
+
+    def test_diff_manifests_err(self):
+        """Diff two valid manifests and verify the output."""
+        # Create two valid manifests, then stunt one in such a way that
+        # `manifest_metaspec` will reject it.
+        valid = make_manifest("Deployment", "namespace", "name1")
+        invalid = make_manifest("Deployment", "namespace", "name2")
+        del invalid["kind"]
+
+        # Test function must return with an error, irrespective of which
+        # manifest was invalid.
+        assert manio.diff_manifests(valid, invalid) == RetVal(None, True)
+        assert manio.diff_manifests(invalid, valid) == RetVal(None, True)
+        assert manio.diff_manifests(invalid, invalid) == RetVal(None, True)
+
 
 
 class TestYamlManifestIOIntegration:
